requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.11
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.6.0
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
starlette==0.46.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
xgboost==3.0.0
